+++ apps/frontend/src/graphql/currencies.graphql.ts
@@ -12,6 +12,7 @@ export const CURRENCY_FRAGMENT = gql`
     symbol
     description
     colorId
+    isTradeable
     archivedAt
     createdAt
     updatedAt